refactor: remove .bind() calls
